go txsource: move workloads to a package
