+++ go/oasis-node/cmd/debug/txsource/workload/transfer.go
@@ -1,4 +1,4 @@
-package txsource
+package workload
 
 import (
 	"context"
@@ -9,6 +9,7 @@ import (
 
 	"github.com/oasislabs/oasis-core/go/common/crypto/signature"
 	memorySigner "github.com/oasislabs/oasis-core/go/common/crypto/signature/signers/memory"
+	"github.com/oasislabs/oasis-core/go/common/logging"
 	"github.com/oasislabs/oasis-core/go/common/quantity"
 	consensus "github.com/oasislabs/oasis-core/go/consensus/api"
 	"github.com/oasislabs/oasis-core/go/consensus/api/transaction"
@@ -21,7 +22,9 @@ const (
 	TransferAmount = 1
 )
 
-func workloadTransfer(rng *rand.Rand, conn *grpc.ClientConn, cnsc consensus.ClientBackend, _ runtimeClient.RuntimeClient) error {
+var logger = logging.GetLogger("cmd/txsource/workload/transfer")
+
+func WorkloadTransfer(rng *rand.Rand, conn *grpc.ClientConn, cnsc consensus.ClientBackend, _ runtimeClient.RuntimeClient) error {
 	// Load all the keys up front. Like, how annoyed would you be if down the line one of them turned out to be
 	// corrupted or something, ya know?
 	accounts := make([]struct {
